[testloop] Move DropCondition from builder to env
