+++ test-loop-tests/src/setup/mod.rs
@@ -1,3 +1,4 @@
 pub mod builder;
+pub mod drop_condition;
 pub mod env;
 pub mod state;
