+++ test-loop-tests/src/setup/state.rs
@@ -25,7 +25,7 @@ use near_store::Store;
 use nearcore::state_sync::StateSyncDumper;
 use tempfile::TempDir;
 
-use super::builder::DropConditionKind;
+use super::drop_condition::DropCondition;
 use super::env::TestLoopChunksStorage;
 
 const NETWORK_DELAY: Duration = Duration::milliseconds(10);
@@ -39,7 +39,7 @@ pub struct SharedState {
     pub network_shared_state: TestLoopNetworkSharedState,
     pub upgrade_schedule: ProtocolUpgradeVotingSchedule,
     pub chunks_storage: Arc<Mutex<TestLoopChunksStorage>>,
-    pub drop_condition_kinds: Vec<DropConditionKind>,
+    pub drop_condition_kinds: Vec<DropCondition>,
     pub load_memtries_for_tracked_shards: bool,
     pub warmup_pending: Arc<AtomicBool>,
 }
